refactor: rename MatrixVector header
Rename MatrixVector.h to MatrixVector.hpp, update its include site and guard, and remove the inherited copyright preamble as requested.

Implemented with Hermes Agent (GPT-5.6-sol).
