--- dmrg/Engine/MatrixVector.h
@@ -1,146 +1,0 @@
-/*
-Copyright (c) 2009, 2026, UT-Battelle, LLC
-All rights reserved
-
-[DMRG++, Version 6+]
-[by G.A., Oak Ridge National Laboratory]
-
-UT Battelle Open Source Software License 11242008
-
-OPEN SOURCE LICENSE
-
-Subject to the conditions of this License, each
-contributor to this software hereby grants, free of
-charge, to any person obtaining a copy of this software
-and associated documentation files (the "Software"), a
-perpetual, worldwide, non-exclusive, no-charge,
-royalty-free, irrevocable copyright license to use, copy,
-modify, merge, publish, distribute, and/or sublicense
-copies of the Software.
-
-1. Redistributions of Software must retain the above
-copyright and license notices, this list of conditions,
-and the following disclaimer.  Changes or modifications
-to, or derivative works of, the Software should be noted
-with comments and the contributor and organization's
-name.
-
-2. Neither the names of UT-Battelle, LLC or the
-Department of Energy nor the names of the Software
-contributors may be used to endorse or promote products
-derived from this software without specific prior written
-permission of UT-Battelle.
-
-3. The software and the end-user documentation included
-with the redistribution, with or without modification,
-must include the following acknowledgment:
-
-"This product includes software produced by UT-Battelle,
-LLC under Contract No. DE-AC05-00OR22725  with the
-Department of Energy."
-
-*********************************************************
-DISCLAIMER
-
-THE SOFTWARE IS SUPPLIED BY THE COPYRIGHT HOLDERS AND
-CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
-WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-COPYRIGHT OWNER, CONTRIBUTORS, UNITED STATES GOVERNMENT,
-OR THE UNITED STATES DEPARTMENT OF ENERGY BE LIABLE FOR
-ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
-OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
-DAMAGE.
-
-NEITHER THE UNITED STATES GOVERNMENT, NOR THE UNITED
-STATES DEPARTMENT OF ENERGY, NOR THE COPYRIGHT OWNER, NOR
-ANY OF THEIR EMPLOYEES, REPRESENTS THAT THE USE OF ANY
-INFORMATION, DATA, APPARATUS, PRODUCT, OR PROCESS
-DISCLOSED WOULD NOT INFRINGE PRIVATELY OWNED RIGHTS.
-
-*********************************************************
-
-*/
-#ifndef DMRG_MATRIX_VECTOR_H
-#define DMRG_MATRIX_VECTOR_H
-
-#include "MatrixVectorKron/MatrixVectorKron.h"
-#include "MatrixVectorOnTheFly.h"
-#include "MatrixVectorStored.h"
-#include <memory>
-
-namespace Dmrg {
-
-template <typename ComplexOrRealType_> class MatrixVector {
-
-public:
-
-	using BaseType                  = MatrixVectorBase<ComplexOrRealType_>;
-	using TypesType                 = typename BaseType::TypesType;
-	using ModelType                 = typename BaseType::ModelType;
-	using ParametersType            = typename ModelType::ParametersType;
-	using ModelHelperType           = typename BaseType::ModelHelperType;
-	using RealType                  = typename BaseType::RealType;
-	using SparseMatrixType          = typename BaseType::SparseMatrixType;
-	using ComplexOrRealType         = typename BaseType::ComplexOrRealType;
-	using value_type                = typename BaseType::value_type;
-	using VectorRealType            = typename BaseType::VectorRealType;
-	using VectorType                = typename BaseType::VectorType;
-	using FullMatrixType            = typename BaseType::FullMatrixType;
-	using HamiltonianConnectionType = typename ModelType::HamiltonianConnectionType;
-	using AuxType                   = typename ModelHelperType::Aux;
-
-	MatrixVector(const ModelType&                 model,
-	             const HamiltonianConnectionType& hc,
-	             const AuxType&                   aux)
-	    : ptr_(create(model, hc, aux))
-	{ }
-
-	SizeType rows() const { return ptr_->rows(); }
-
-	SizeType cols() const { return ptr_->cols(); }
-
-	void matrixVectorProduct(VectorType& x, const VectorType& y) const
-	{
-		ptr_->matrixVectorProduct(x, y);
-	}
-
-	void fullDiag(VectorRealType& eigs, FullMatrixType& fm) const { ptr_->fullDiag(eigs, fm); }
-
-	const SparseMatrixType& toCRS() const { return ptr_->toCRS(); }
-
-private:
-
-	static std::unique_ptr<BaseType>
-	create(const ModelType& model, const HamiltonianConnectionType& hc, const AuxType& aux)
-	{
-		if (model.params().options.isSet("MatrixVectorStored"))
-			return std::make_unique<MatrixVectorStored<ComplexOrRealType>>(
-			    model, hc, aux);
-
-		if (model.params().matrix_solver_enum
-		    == ParametersType::MatrixSolverEnum::ARNOLDISAI) {
-			throw PsimagLite::RuntimeError(
-			    "MatrixSolver=ArnoldiSaI requires MatrixVectorStored\n");
-		}
-
-		if (model.params().options.isSet("MatrixVectorOnTheFly"))
-			return std::make_unique<MatrixVectorOnTheFly<ComplexOrRealType>>(
-			    model, hc, aux);
-
-		return std::make_unique<MatrixVectorKron<ComplexOrRealType>>(model, hc, aux);
-	}
-
-	std::unique_ptr<BaseType> ptr_;
-};
-
-} // namespace Dmrg
-
-#endif // DMRG_MATRIX_VECTOR_H
